Update and rename UserContext.js to user-context.js
