--- UserContext.js
@@ -1,54 +1,0 @@
-import { useState, useEffect, createContext, useContext } from "react";
-import {NotificationManager} from "react-notifications"
-import { BackendApi } from "../BackendApi"
-
-
-const UserContext = createContext({
-    user: null,  //When No user is logged in
-    loginUser: () =>{},
-})
-
-const useUser = ()=>useContext(UserContext)
-
-const UserProvider =({children}) =>{
-    const [user, setUser] = useState(null);
-    const [isAdmin, setIsAdmin] = useState(false);
-
-    useEffect(() =>{
-        setIsAdmin (user && user.role =='admin')},
-    [user])
-    useEffect(() =>{
-      BackendApi.user.getProfile().then(({user, error}) =>{
-        if(error){
-            console.error(error)
-        }
-        else{
-            setUser(user)
-        }
-      } )
-    },[]).catch(console.error)
-
-    const loginUser = async (username, password) => {
-        const { user, error } = await BackendApi.user.login(username, password)
-        if (error) {
-            NotificationManager.error(error)
-        } else {
-            NotificationManager.success("Logged in successfully")
-            setUser(user)
-        }
-    }
-
-    const logoutUser = async () => {
-        setUser(null)
-        //Delete the maintained session
-        await BackendApi.user.logout()
-    }
-    
-    //wrappped 
-    return (
-        <UserContext.Provider value={{ user, loginUser, logoutUser, isAdmin }}>
-            {children}
-        </UserContext.Provider>
-    )
-}
-export {useUser, UserProvider}
